@@ -3,24 +3,25 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@react-three/drei": "^7.4.0",
-    "@react-three/fiber": "^7.0.6",
-    "@react-three/xr": "^3.1.2",
+    "@react-three/drei": "^9.68.3",
+    "@react-three/fiber": "^8.13.0",
+    "@react-three/xr": "^5.5.0",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
+    "@types/three": "^0.152.0",
     "axios": "^0.21.4",
     "bootstrap": "^4.6.0",
     "formik": "^2.2.9",
     "http-proxy-middleware": "^2.0.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^1.6.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-google-recaptcha": "^2.1.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
-    "three": "^0.131.3",
+    "three": "^0.152.2",
     "web-vitals": "^1.1.2",
     "yup": "^0.32.9"
   },
